Documentation of Project
Over the last decades machine learning has become very popular since new algorithms were developed and more complex implementations of neural networks became possible, opening a new world of possibilities and solutions. Nowdays is very common to use tools and software that already have machine learning algorithms implemented, and people only need to know the necessary theory to acomplish systems that learn and do what they need. Still, when working with machine learning it is very important to understand the basics and go deeper in the algorithms so the student gets a better understanding of this mathematical tools and later he can get more involved with AI without much trouble. This is why we decided to implement a Shallow Neural Network from scratch in C++ that learns how to classify handwritten digits with 84.42% precision. We did not use any machine learning library and everything was implemented in C++.